+++ lib/app_store_server_library/models/subscription_group_identifier_item.ex
@@ -7,6 +7,12 @@ defmodule AppStoreServerLibrary.Models.SubscriptionGroupIdentifierItem do
 
   alias AppStoreServerLibrary.Models.LastTransactionsItem
 
+  @doc false
+  @spec __nested_fields__() :: %{atom() => {:list | :single, module()}}
+  def __nested_fields__ do
+    %{last_transactions: {:list, LastTransactionsItem}}
+  end
+
   @type t :: %__MODULE__{
           subscription_group_identifier: String.t() | nil,
           last_transactions: [LastTransactionsItem.t()] | nil